Add basic ansible configuration and documentation
Ensure the ansible-vault password is added at `~/ansible-password.txt`
so that it can decrypt the vault at runtime.

Host `pve` uses 2fa authentication, which means that a `ControlPath`
socket for an authenticated connection needs to be dumped for ansible
to use. So, before any `ansible` commands can be run. Configure
`~/.ssh/config` and run `ssh pve`.
```bash
Host pve
	Hostname 192.168._.__
	User ________
	IdentityFile ~/.ssh/__________.key
	SetEnv TERM="xterm-256color"
	Port ____
	# enable ssh multiplexing
	ControlPath /tmp/%r@%h:%p
	ControlMaster auto
	ControlPersist 30m
```