Update group name and documentation
The group name was changed from `servers` because it was deemed too
generic and potentially a cause of confusion when the inventory
expands.

The ansible documentation had to be updated to reflect such a change,
and in the process a few minor problems such as a broken footnote and
a lengthy paragraph were resolved.