Add sub-playbooks and use `--limit` cli option
I use sub-playbooks and not tasks, so to get approved by the linter
every playbook must have a `hosts:` value.

But, when I want to target a specific host like 'pve' by modifying
`main.yml` to `hosts: pve`. It isn't actually respected because all the
sub-playbooks have `hosts: all`.

Which is why I've resorted to using the cli option `--limit`