Fix crash when `locale_language` isn't found
When there is no `locale_language` value it defaults to `locale_lang`
but it seems that a variable is only defined and exposed after the task
is completed.

So, if both variables were put under the same task, when
`locale_language` needs to default. `locale_lang` isn't actually defined
yet and would cause a crash.