Add a playbook to optimize sysctl/services/hugepages
I needed to guarantee that the tasks `Load sysctl settings from files`
and `Run update-grub command` only ran if the task before has changed.

Which would normally mean creating a handler instead. But, then I had
to figure out someway to ensure that the handler `Reboot System` only
ever ran after all other handlers had finished.

This would mean extracting the handlers out of the playbooks themselves
and using roles because "Notify handlers are always run in the same
order that they are defined, not in the order listed in the notify
statement" [^1].

Which meant I would have to rewrite all current playbooks as roles. I
did make a start but then remembered that trying to play with roles is
what killed one of my earlier ansible attempts. So, I've decided to just
ignore the `ansible-lint` warnings.

[^1]: https://stackoverflow.com/q/35130051